Added additional option to not automatically delete new "Thread Notification"

+++ src/events/server/threadCreate.ts
@@ -21,9 +21,12 @@ const event: Event = {
 				lastMessage?.system &&
 				lastMessage.author.id === client.user!.id
 			) {
-				if ((record && record.SupportChannel === parentId) || name.toLowerCase() === 'support') {
-					lastMessage.delete();
+				if (!record.ThreadNotification) {
+					if ((record && record.SupportChannel === parentId) || name.toLowerCase() === 'support') {
+						lastMessage.delete();
+					}
 				}
+
 			}
 		} catch (err) {
 			console.error(err);
